refactor: integrate DTO and exception handling in user removal
- Updated IUserService interface to include Response<String> for user removal
- Refactored deleteUser method to check for existence before deletion and throw ResourceNotFoundException if not found
- Implemented logging for successful book deletion operations

+++ src/main/java/com/example/Librarymanagementsystem/service/impl/UserService.java
@@ -2,12 +2,16 @@ package com.example.Librarymanagementsystem.service.impl;
 
 import com.example.Librarymanagementsystem.data.model.User;
 import com.example.Librarymanagementsystem.data.repository.UserRepository;
+import com.example.Librarymanagementsystem.exception.ResourceNotFoundException;
+import com.example.Librarymanagementsystem.payload.response.Response;
 import com.example.Librarymanagementsystem.service.IUserService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -43,7 +47,19 @@ public class UserService implements IUserService {
         return userRepository.save(user);
     }
 
-    public void deleteUser(Long id) {
+
+    public Response<String> deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            log.warn("User with ID {} not found", id);
+            throw new ResourceNotFoundException("User with ID " + id + " not found.");
+        }
         userRepository.deleteById(id);
+        log.info("User deleted successfully with ID: {}", id);
+
+        return new Response<>(
+                LocalDateTime.now(),
+                HttpStatus.OK.value(),
+                "User deleted successfully with ID: " +id
+        );
     }
 }
